DTO for Vehicle created
